refactor(activity): move dungeon score types to lib layer
The `CharacterDungeonScores` and `DungeonRunSummary` interfaces lived in
`services/raiderioMythicPlus.ts` but were consumed by `lib/dungeonSuggestions.ts`,
inverting the intended layer dependency: `services/` is the I/O boundary that
should depend on `lib/`, not vice versa.

Moved both interfaces to a new `lib/dungeonScoreTypes.ts`. The runtime
function `fetchCharacterDungeonScores` stays in `services/`. To preserve
backward compatibility, `services/raiderioMythicPlus.ts` now re-exports
the types so any external consumer importing from there keeps working.

Updated import statements in:
- activity/src/lib/dungeonSuggestions.ts
- activity/src/lib/dungeonSuggestions.test.ts
- activity/src/hooks/useDungeonSuggestions.ts
- activity/src/components/GroupCarousel.tsx
- activity/src/components/GroupSlide.tsx
- activity/src/components/SpotlightPortrait.tsx
- activity/src/components/SpotlightPortrait.stories.tsx
- activity/src/components/SpotlightPortraits.tsx

Pure type-only move; no runtime change.
