+++ activity/src/services/raiderioMythicPlus.ts
@@ -1,36 +1,11 @@
 // Public Raider.io endpoints — same hostname/contract used by `lookupCharacterProfile`.
 // CORS is enabled, so we can call directly from the browser.
 
-export interface DungeonRunSummary {
-  /** Stable identifier (challenge_mode_id) for matching across runs and seasons. */
-  challengeModeId: number;
-  /** Display name (e.g. "Operation: Floodgate"). */
-  name: string;
-  /** Short tag (e.g. "FLOOD"). */
-  shortName: string;
-  /** Best key level the character has timed/run for this dungeon, or 0. */
-  level: number;
-  /** Raider.io score for the best run, or 0. */
-  score: number;
-  /** Dungeon icon URL from Raider.io's CDN, or null when missing. */
-  iconUrl: string | null;
-}
+import type { CharacterDungeonScores, DungeonRunSummary } from '../lib/dungeonScoreTypes';
 
-export interface CharacterDungeonScores {
-  name: string;
-  realm: string;
-  region: string;
-  /** Best run per dungeon, keyed by challenge_mode_id. */
-  byDungeon: Record<number, DungeonRunSummary>;
-  /** Overall Raider.io M+ score for the current season ("all" segment), or null. */
-  overallScore: number | null;
-  /** Raider.io rarity color for the overall score (e.g. "#f06862"), or null. */
-  scoreColor: string | null;
-  /** Character's class name from Raider.io (e.g. "Mage"), or null. */
-  className: string | null;
-  /** Active spec name (e.g. "Frost"), or null. */
-  specName: string | null;
-}
+// Re-export the type contracts so existing consumers that import from this
+// module keep working. The canonical home for the types is `lib/dungeonScoreTypes`.
+export type { CharacterDungeonScores, DungeonRunSummary } from '../lib/dungeonScoreTypes';
 
 interface RaiderioRun {
   dungeon: string;
